+++ common/analysis/InsertQueryAnalyser.java
@@ -16,7 +16,7 @@
  *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package grakn.benchmark.profiler.analysis;
+package grakn.benchmark.common.analysis;
 
 
 import grakn.core.concept.Concept;
