+++ common/analysis/test/InsertQueryAnalyserTest.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/agpl.txt>.
  */
 
-package grakn.benchmark.profiler.analysis;
+package grakn.benchmark.common.analysis;
 
 import grakn.core.concept.Concept;
 import grakn.core.concept.ConceptId;
