+++ common/configuration/test/BUILD
@@ -1,0 +1,10 @@
+java_test(
+    name = "configuration-test",
+    test_class = "grakn.benchmark.common.configuration.ConfigurationTest",
+    srcs = ["ConfigurationTest.java"],
+    deps = [
+        "//common/configuration:benchmark-configuration",
+        "//profiler/src:benchmark-profiler",
+        "//dependencies/maven/artifacts/commons-cli",
+    ],
+)
