+++ generator/test/provider/BUILD
@@ -1,6 +1,6 @@
 java_test(
     name = "central-stream-provider-test",
-    test_class = "grakn.benchmark.profiler.generator.provider.concept.CentralConceptProviderTest",
+    test_class = "grakn.benchmark.generator.provider.concept.CentralConceptProviderTest",
     srcs = ["CentralConceptProviderTest.java"],
     deps = [
         "@graknlabs_grakn_core//concept:concept",
@@ -11,7 +11,7 @@ java_test(
 
 java_test(
     name = "conceptid-storage-provider-test",
-    test_class = "grakn.benchmark.profiler.generator.provider.concept.ConceptIdStorageProviderTest",
+    test_class = "grakn.benchmark.generator.provider.concept.ConceptIdStorageProviderTest",
     srcs = ["ConceptIdStorageProviderTest.java"],
     deps = [
         "@graknlabs_grakn_core//concept:concept",
@@ -22,7 +22,7 @@ java_test(
 
 java_test(
      name = "not-in-relationship-conceptid-storage-provider-test",
-     test_class = "grakn.benchmark.profiler.generator.provider.concept.NotInRelationshipConceptIdProviderTest",
+     test_class = "grakn.benchmark.generator.provider.concept.NotInRelationshipConceptIdProviderTest",
      srcs = ["NotInRelationshipConceptIdProviderTest.java"],
      deps = [
         "@graknlabs_grakn_core//concept:concept",
