+++ generator/test/query/BUILD
@@ -1,6 +1,6 @@
 java_test(
     name = "attribute-generator-test",
-    test_class = "grakn.benchmark.profiler.generator.query.AttributeGeneratorTest",
+    test_class = "grakn.benchmark.generator.query.AttributeGeneratorTest",
     srcs = ["AttributeGeneratorTest.java"],
     deps = [
         "@graknlabs_graql//java:graql",
@@ -12,7 +12,7 @@ java_test(
 
 java_test(
     name = "entity-generator-test",
-    test_class = "grakn.benchmark.profiler.generator.query.EntityGeneratorTest",
+    test_class = "grakn.benchmark.generator.query.EntityGeneratorTest",
     srcs = ["EntityGeneratorTest.java"],
     deps = [
         "@graknlabs_graql//java:graql",
@@ -24,7 +24,7 @@ java_test(
 
 java_test(
     name = "relationship-generator-test",
-    test_class = "grakn.benchmark.profiler.generator.query.RelationGeneratorTest",
+    test_class = "grakn.benchmark.generator.query.RelationGeneratorTest",
     srcs = ["RelationGeneratorTest.java"],
     deps = [
         "@graknlabs_graql//java:graql",
