+++ profiler/src/GraknBenchmark.java
@@ -29,8 +29,9 @@ import grakn.benchmark.generator.query.QueryProvider;
 import grakn.benchmark.generator.storage.ConceptStorage;
 import grakn.benchmark.generator.storage.IgniteConceptStorage;
 import grakn.benchmark.generator.util.IgniteManager;
-import grakn.benchmark.profiler.util.BenchmarkArguments;
-import grakn.benchmark.profiler.util.BenchmarkConfiguration;
+import grakn.benchmark.common.configuration.BenchmarkArguments;
+import grakn.benchmark.common.configuration.BenchmarkConfiguration;
+import grakn.benchmark.profiler.util.BootupException;
 import grakn.benchmark.profiler.util.ElasticSearchManager;
 import grakn.benchmark.profiler.util.SchemaManager;
 import grakn.benchmark.profiler.util.TracingGraknClient;
