+++ profiler/src/util/BootupException.java
@@ -1,6 +1,6 @@
-package grakn.benchmark.profiler;
+package grakn.benchmark.profiler.util;
 
-public class BootupException extends RuntimeException{
+public class BootupException extends RuntimeException {
     public BootupException(String message, Throwable cause) {
         super(message, cause, false, false);
     }
