+++ profiler/test/BUILD
@@ -4,6 +4,7 @@ java_test(
     srcs = ["ProfilerBootupTestIntegration.java"],
     deps = [
         "//profiler/src:benchmark-profiler",
+        "//common/configuration:benchmark-configuration",
         "@graknlabs_client_java//:client-java",
         "@graknlabs_grakn_core//concept:concept",
         "@graknlabs_graql//java:graql",
@@ -15,16 +16,3 @@ java_test(
     ]
 )
 
-java_test(
-    name = "profiler-bootup-test",
-    test_class = "grakn.benchmark.profiler.ProfilerBootupTest",
-    srcs = ["ProfilerBootupTest.java"],
-    deps = [
-        "//profiler/src:benchmark-profiler",
-        "//dependencies/maven/artifacts/commons-cli",
-    ],
-    classpath_resources = ["//profiler/test/resources:logback-test"],
-    data = [
-        "//profiler/test/resources:web-content-config"
-    ]
-)
