+++ report/BUILD
@@ -1,0 +1,3 @@
+java_library(
+    name = "report"
+)
